refactor:  기존 profile/edlit 디렉토리  하위 컴포넌트 이름 수정 및 변수 명 수정
