+++ src/app/profile/page.tsx
@@ -1,16 +1,16 @@
 import { getProfile } from "@/app/auth/lib/actions";
-import ProfileEdit from "./edit/page";
+import ProfileEdit from "./edit/ProfileEditWrapper";
 import styles from '@/app/profile/profile.module.css';
 
 export default async function ProfilePage() {
-  const profile = await getProfile()
+  const profile = await getProfile() || { image_url: '', name: '', id: '' };
   return (
     <div className={styles.profileContainer}>
       <h1 className={styles.title}>프로필 수정</h1>
       <ProfileEdit
-        initialImageUrl={profile?.image_url}
-        initialNickname={profile?.name}
-        initialId={profile?.id}
+        image_url={profile?.image_url}
+        name={profile?.name}
+        id={profile?.id}
       />
     </div>
   );
